% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_risk_assessment.R
\name{create_risk_assessment}
\alias{create_risk_assessment}
\title{FHIR RiskAssessment}
\usage{
create_risk_assessment(assessment_date, patient_id, risk)
}
\arguments{
\item{assessment_date}{date.time, when was the assessment made}

\item{patient_id}{The patient id}

\item{risk}{Predicted Risk for Major Ampuation}
}
\value{
A FHIR RiskAssessment Resource (json)
}
\description{
Creats a FHIR RiskAssessment Resource
}
\examples{
assessment_date <- Sys.time()
patient_id <- 123
risk <- 1 / 10
create_risk_assessment(assessment_date, patient_id, risk)
}
