% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_risk_assessment.R
\name{generate_risk_assessment}
\alias{generate_risk_assessment}
\title{Returns JSON Template
Template for the FHIR RiskAssessment Resource
Risk values has to be entered}
\usage{
generate_risk_assessment()
}
\value{
A parsed JSON
}
\description{
Returns JSON Template
Template for the FHIR RiskAssessment Resource
Risk values has to be entered
}
