% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-histogram.R
\name{histogramm_posterior}
\alias{histogramm_posterior}
\title{Return Posterio Historgramm for given}
\usage{
histogramm_posterior(pedis, parameter = param_dist)
}
\arguments{
\item{pedis}{The PEDIS Classes in a Vector}

\item{parameter}{The posterior distribution (a matrix n times 7)}
}
\value{

}
\description{
Return Posterio Historgramm for given
}
\examples{
pedis <- c("P" = 3, "E" = 3, "D" = 3, "I" = 1, "S" = 1, "Age" = 60, "Gender" = 1) # Gender Male

}
