% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Initialize the Model}
\usage{
model(
  data = pedis,
  formula = any_amputation ~ p + e_ordinal_5 + d + i + s + alter_bei_aufnahme + gender,
  prior = cauchy(scale = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5)),
  n_iter = 1000,
  thinning = NULL,
  seed = 234241,
  ...
)
}
\value{

}
\description{
Initialize the Model
}
