% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{risk_statistic}
\alias{risk_statistic}
\title{Predict Amputation Risk}
\usage{
risk_statistic(
  perfusion = 3,
  extent = 3,
  depth = 1,
  infection = 1,
  sensation = 2,
  age = 1,
  gender = 1
)
}
\arguments{
\item{perfusion}{int: 1-3}

\item{extent}{int: 1-3}

\item{depth}{int: 1-3}

\item{infection}{int: 1-4}

\item{sensation}{int: 1-2}

\item{age}{int}

\item{gender}{int: 0-male, 1-male}
}
\value{
The estimated probability
}
\description{
Predict Amputation Risk
}
\examples{
p <- predict()
mean(p)
}
